Restructuring: separate index with it's disk state and index subtree restructing
