+++ src/blob/mod.rs
@@ -12,6 +12,6 @@ pub(crate) use super::prelude::*;
 
 mod prelude {
     pub(crate) use super::*;
-    pub(crate) use index::Simple as SimpleIndex;
+    pub(crate) use index::Index;
     pub(crate) use std::sync::atomic::AtomicU64;
 }
